backend/src/main/kotlin/ase/athlete_view/domain/fitness/service/FitnessService.kt
/*-
 * #%L
 * athlete_view
 * %%
 * Copyright (C) 2023 - 2024 TU Wien INSO ASE GROUP 5 WS2023
 * %%
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 * 
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 * 
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 * #L%
 */
package ase.athlete_view.domain.fitness.service

import ase.athlete_view.common.exception.entity.ForbiddenException

interface FitnessService {
    /**
     * Calculates the fitness of a target user.
     * Fitness is calculated by Health and Activity. If the Activity is not provided -> this day has less fitness.
     * The optimal fitness is between 70 and 90 %.
     *
     * @param byUser is who asks the data. Only the same athlete or his/her trainer can access this data.
     * @param targetUser is who the data belongs to.
     * @return the calculated fitness.
     * @throws ForbiddenException when user byUser has no access or targetUser is a trainer.
     */
    fun calculate(byUser: Long, targetUser: Long): List<Int>
}
